sssss

+++ src/pages/DashBoard/AllOders/AllOders.js
@@ -19,29 +19,16 @@ const AllOders = () => {
     }
 
     const handleDelterOders = (id) => {
-        fetch(`https://frozen-anchorage-07301.herokuapp.com/delteoders/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json)
-            .then(data => {
-                console.log(data);
+        const isDelete = window.confirm('are you sure delete');
+        if (isDelete) {
+            fetch(`https://frozen-anchorage-07301.herokuapp.com/delteoders/${id}`, {
+                method: 'DELETE'
             })
-        swal({
-            title: "Are you sure?",
-            text: "Once deleted, you will not be able to recover this imaginary file!",
-            icon: "warning",
-            buttons: true,
-            dangerMode: true,
-        })
-            .then((willDelete) => {
-                if (willDelete) {
-                    swal("Poof! Your imaginary file has been deleted!", {
-                        icon: "success",
-                    });
-                } else {
-                    swal("Your imaginary file is safe!");
-                }
-            });
+                .then(res => res.json)
+                .then(data => {
+                    console.log(data);
+                })
+        }
     }
     return (
         <div className="">
